==> Notebooks/1.1 - Collecting journalist data.ipynb <==
### Load libaries and files

import pickle
import pandas as pd
import json
import langdetect
import re
import Algorithmia
import bs4 as bs
import lxml
from urllib.request import Request, urlopen
import time
from fuzzywuzzy import process
import tfidf_matcher as tm
from fake_useragent import UserAgent

with open('../Data/pickled_df.pkl', 'rb') as f:
    data = pickle.load(f)
data = data.reset_index(drop = True)
muckrack_sitemap = pd.read_csv('../Data/muckrack_persons_fetchlist.csv')

data

### Initial preprocessing

#### Remove unnecessary columns

data = data.drop(['crawl_date', 'crawl_desc', 'crawl_url', 'img_url', 'webpage_source', 'lang'], 1)

data

#### Convert lists to individual journalists

articles = []
for j in range(len(data)):
    try:
        contributors = json.loads(data.iloc[j].author)
        for contributor in contributors:
            articles.append([contributor, data.iloc[j].date, data.iloc[j].description, data.iloc[j].full_text, 
                         data.iloc[j].site_name, data.iloc[j].title, data.iloc[j].topic, data.iloc[j].url])
    except:
        pass

articles = pd.DataFrame(articles)
articles.columns = ['author', 'date', 'description', 'full_text', 'site_name', 'title', 'topic', 'url']

articles

#### Remove outlets / organisations from list of author names

remove = {'', 'Reuters', 'Reuters Staff', 'Associated', 'Vanguard', 'ABC News', 
          'The Washington Times http://www.washingtontimes.com', 'BWW News Desk', 'Press Trust of India', 
          'Press Association', 'PTI', 'Capital Market', 'PA Media', 'The Korea Herald', 'SA Transcripts',
          'Deutsche Welle (www.dw.com)', 'Associated Press', 'Afp', 'India.com News Desk', 'The Associated Press',
          'Australian Associated', 'FE Online', 'www.ETAuto.com', 'www.ETTelecom.com', 'F_488', 'Staff Writer',
          'India.com Sports Desk', 'www.ETEnergyworld.com', 'IANS', 'TV News Desk', 'ANI Press Release', 'AP', 
          'www.ETHealthworld.com', 'WRAL', 'India.com Entertainment Desk', 'ANI', 'HT Auto Desk', 
          'WND News Services', 'Stage Tube', 'Staff report', 'The Economist', 'AFP', 'F_300926', 
          'www.ETBrandEquity.com', 'FE Bureau', 'The Christian Science Monitor', 'Flanders News', 'BS Web Team',
          'India.com Lifestyle Staff', 'VietNamNet News', 'India.com Lifestyle Staff', 'India.com Viral News Desk',
          'Dailymail.com', 'The Moscow Times', 'www.ETCIO.com', 'www.ETCIO.com', 'SI Reporter', 
          'From Staff Reports','BWW', 'Entrepreneur en Español', 'Trefis Team', 'azfamily.com News Staff', 
          'From staff reports', 'Compiled by Democrat-Gazette Staff From Wire Reports', 
          'Classifieds Arizona Daily Star', '', 'Daily Express newspaper', 'www.ETRealty.com', 'Daily', 'Livemint',
          'Kyodo', 'India.com Business Desk', 'The Editorial Board', 'F_300833', 'HNN Staff','F_126', 'F_127',
          'India.com Education Desk', 'F_300824', 'Staff', 'doctor.ndtv.com', 'Daily News Editorial Board',
          'Expert Panel®', 'Staff Report', 'Entrepreneur Store',  'BWW Staff', 'to the editor', 'Dawn.com', 
          'WND Staff', 'Arizona Daily Star',  'F_300832', 'TV Scoop', None, 'STLNEWS', 
          'India.com Hindi News Desk', 'news18', "The Newspaper's Staff Reporter", 
          'Business Standard Editorial Comment', 'Business Standard', 'THE ASSOCIATED PRESS',  'Associated Press staff',
          'THE EDITORIAL STAFF', 'The Value Portfolio', 'Breakingviews columnists', 'StatBot | STLhighschoolSPORTS.com',
          'Team YS', 'F_200745', 'F_300923', 'Special to The Commercial', 'Tarot Astrologers',
          'Voice of the People', 'TNW Deals', 'Guardian community team', 'Baltimore Sun staff', 'News18',
          'Mail', 'AP Reporters', 'Avisol Capital Partners', 'Agencies', 'News staff', 'CNBC.com staff',
          'BroadwayWorld TV', 'The Conversation', 'F_200798', ' birminghammail.co.uk', 'ASSOCIATED PRESS',
          'F_200805', "The Age's View", 'BOOX Research', 'Associated Press Staff', 'Fun Trading', 'The Gazette',
          'American Heart Association News', 'Record Reporter', 'The Investment Doctor', 'Roadshow staff',
          'the Times Union Editorial Board', 'Baltimore Sun Editorial Board', 'The Editors of\xa0GQ', 
          'The Roanoke Times', 'The Spectator', 'Editorial staff', 'Individual Trader', 'Univision', None,
          'Arkansas Democrat-Gazette', 'Mirror.co.uk', 'ASSOCIATED PRESS STAFF', 'DerWesten - derwesten.de',
          'The New York Times', 'Wolf Report', 'The Gazette editorial board', 'Real Vision', 
          'Sun Sentinel Editorial Board', 'Voice of the Mirror', '@adwait', 'Expert Panel', 'The Buy Area Staff',
          'Volatility Surfer', 'Special To The Arizona Daily Star', 'NFL Staff', 'CNN Staff',
          'PJ Brown Special to the Arizona Daily Star', 'Invesco US', 'Bloomberg News', 'Quora', 'CNN Wire',
          'IP Banking Research', 'Daile Cross', 'The Washington Post', 'Bears of Wall Street', 'Teen Vogue Staff',
          'ICICI Securities', 'WIRED Staff', 'This Is', 'Cinemanía', 'STAFF EDITORIAL', 'StackCommerce', 'Axios',
         'Lincoln Journal Star', 'The Sunday Times', 'IE Online', 'Bloomberg', 'Associated Press Reporter' ,'Behance',
         'The Financial Express', 'Firstpost', '20minutos', 'Daily Herald report', 'Staff reports', 'MarketWatch',
         'Times Staff', 'Trapping Value', 'Q.ai - Investing Reimagined', 'CNET staff', 'YEC', 'Catholic Online', 
         'Stone Fox Capital', 'PA Sport Staff', 'Entrepreneur Staff', 'Keystone-SDA/ts', 'Associated Press reporters',
         "The Herald's View", 'Autoblog Staff', 'finanzen.net GmbH', 'Rolling Stone', 'AP | PTI', 'Story by Reuters',
         'PA reporters', 'Sports Hotline', 'Pioneer Press Staff', 'PA', 'Ticked Off!', 'Gazette readers', 'ValueZen',
         'India Knight', 'RMC SPORT', 'Truthout', 'MailOnline', 'Variety Staff', 'News Service of Florida', 'From wire reports',
         'Elephant Analytics', 'Chowhound Staff', 'Quad 7 Capital', 'The Associated Press, Associated Press',
         'undefined undefined', 'Orlando Sentinel Editorial Board', 'Clarín.com', 'Staff and wire reports', 'IndiaToday.in',
         'East Kilbride News', 'Retirement Pot', 'QuandaryFX', 'Perthshire Advertiser', 'The Inquirer Editorial Board',
         'CondÃ© Nast', 'WRITE TO LETTERS@THETIMES.CO.UK', 'Entrepreneur Middle East Staff', 'Editorial', 'WWD Staff',
         'STATE JOURNAL WIRE SERVICES', 'Think Change India', 'Shock Exchange', 'News4Jax staff', 'CNN staff',
         'The A.V. Club', 'Journal Star editorial board', 'Record View', 'Renaissance Capital IPO Research', 
          'MarketWatch Automation', 'Morning Call Staff', 'Pa Sport Staff', 'The Value Investor', 'Casa Editorial El Tiempo',
         'Hamilton Advertiser', 'HELLO! Canada', 'The Virginian-Pilot & Daily Press Editorial Board', 'Wishaw Press',
          'By CNN Underscored Staff', 'Political Comments', 'Mott Capital Management', 'Captain Alpha', 'STATE JOURNAL STAFF',
          'By <a href="/profiles/meg-wagner">Meg Wagner</a> and <a href="/profiles/melissa-mahtani">Melissa Mahtani</a>, CNN',
          'Franklin Templeton Investments', 'Ellevate', 'Pa', 'PA Reporters', 'Chicago Sun-Times', 'Navbharat Times', 'Long Player', 
          'By <a href="/profiles/meg-wagner">Meg Wagner</a>, Melissa Macaya, <a href="/profiles/melissa-mahtani">Melissa Mahtani</a>, <a href="/profiles/veronica-rocha">Veronica Rocha</a> and <a href="/profiles/amanda-wills">Amanda Wills</a>, CNN',
          'By <a href="/profiles/meg-wagner">Meg Wagner</a>, Melissa Macaya and Mike Hayes, CNN', 'SAPO', 'Slate Staff',
          'By <a href="/profiles/meg-wagner">Meg Wagner</a>, <a href="/profiles/fernando-alfonso-iii">Fernando Alfonso III</a>, Melissa Macaya, <a href="/profiles/melissa-mahtani">Melissa Mahtani</a>, <a href="/profiles/veronica-rocha">Veronica Rocha</a> and <a href="/profiles/amanda-wills">Amanda Wills</a>, CNN', 
          'By <a href="/profiles/donie-osullivan">Donie ' + "O'Sullivan</a>, " + '<a href="/profiles/kaya-yurieff">Kaya Yurieff</a>, Kelly Bourdet, the CNN Business team and contributors from across CNN', 
          'Japan Today', 'BrandWagon Online', 'Condé Nast', 'Gen Alpha', 'Chicago Tribune staff', 'Field Level Media', 
         'Tribune Content Agency', 'Opal Investment Research', 'Steerpike', 'The Value Pendulum', 'Sheen Bay Research',
         'SAP Guest', 'Guardian sport', 'Value Kicker', 'Staff Reports', 'Dividend Seeker', 'Wall Street Breakfast',
         'Stackcommerce', 'Keystone-SDA/ac', 'Compiled by Democrat-Gazette staff from wire reports', 'Markit', 'WisdomTree',
         'Carroll County Times Editorial Board', 'Hoya Capital Real Estate', 'NWA Democrat-Gazette'}

# remove other unicodes
for i in range(len(articles)):
    try:
        if not ''.join(filter(lambda character:ord(character) < 0x3000, articles.iloc[i].author)):
            remove.add(articles.iloc[i].author)
    except:
        pass
        
articles = articles[~articles['author'].isin(remove)]

articles

#### Remove journalists who have authored less than 10 articles

articles['author_article_count'] = articles.groupby('author')['author'].transform('count')
articles = articles[articles.author_article_count >= 10]

articles = articles.reset_index(drop = True)

articles

#### Remove irrelevant keywords from name

def format_name(name):
    clean_name = name.lower()
    clean_name = re.sub(r'\|.*$', '', clean_name) # remove everything after |
    clean_name = re.sub(r'\,.*$', '', clean_name) # remove evrything after ,
    punctuations = '''!()[]{};:'"\,<>./?#$%^&*_~'''
    
    for x in clean_name.lower(): 
        if x in punctuations: clean_name = clean_name.replace(x, "") # remove punctuations except hyphen
    clean_name = re.sub(r'(\d+)', '', clean_name) # remove numbers
    clean_name = re.sub(r'(--)', '', clean_name)
    clean_name = ' '.join([i.strip() for i in clean_name.split()]) # remove spaces and lowercase
    clean_name = re.sub(r'(\w*@\w*)', '', clean_name) # remove emails
        
    clean_name = re.sub(r'(sa)(\s+)(\w+)(\s+)', '', clean_name) # SA editors
    clean_name = re.sub(r'(tulsa)(\s+)(world)', '', clean_name) # tulsa world editors
    clean_name = re.sub(r'(world-herald)(\s*)(\w*)(\s*)(\w*)', '', clean_name) # world herald editors
    clean_name = re.sub(r'(news)(\s+)(editor)', '', clean_name) # news editors
    clean_name = re.sub(r'(richmond)(\s+)(times-dispatch)', '', clean_name) # richmond editors
    clean_name = re.sub(r'(new)(\s+)(hampshire)(\s+)(union)(\s+)(leader)', '', clean_name) # new hampshire editors
    clean_name = re.sub(r'(for)(\s+)(the)(\s+)(state)(\s+)(journal)', '', clean_name) # state journal editors
    clean_name = re.sub(r'(union)(\s+)(leader)(\s+)(correspondent)', '', clean_name) # union leader correspondants
    clean_name = re.sub(r'(sfgate)', '', clean_name) # sfgate correspondants
    clean_name = re.sub(r'(special)(\s+)(from)(\s+)(the)(\s+)(gazette)', '', clean_name) # the gazette correspondants
    clean_name = re.sub(r'(times)(\s+)(correspondent)', '', clean_name) # times correspondants
    
    keywords = ['sa', 'tulsa', 'editor', 'writer', 'world-herald', 'news', 'world', 'richmond', 'times-dispatch', 'new',
                   'hampshire', 'union', 'leader', 'for', 'the', 'state', 'journal', 'correspondant', 'sfgate', 'special',
                   'from', 'the', 'gazette', 'times', 'staff', 'senior', 'dr', 'correspondent', 'by', 'editorial board',
                   'research', 'wire reports', 'security', 'real', 'estate', 'to', 'post', 'and', 'courier', 'policy',
                   'commercial', 'bureau', 'political', 'roanoke', 'college', 'football', 'editorial','democrat-gazette',
                   'arizona', 'daily', 'star', '--hamburg', 'column', 'lincoln', 'managing', 'backstage', 'with', 'sports',
                   'ii', 'iii', 'capitol', 'media', 'services']
    clean_name = ' '.join([i for i in clean_name.split() if i not in keywords]) # remove other keywords
    clean_name = re.sub(r'(@)', '', clean_name) # remove handles

    return clean_name.title()

articles['author_name_clean'] = articles.apply(lambda x: format_name(x['author']), axis=1)

articles

journalists = articles.copy(deep=True)[['author_name_clean', 'author_article_count']].drop_duplicates().reset_index(drop = True)

journalists

### Gathering Muckrack URLs

#### Creating raw URLs

def create_url(name):
    url = 'https://muckrack.com/' + name.lower().replace(' ', '-').strip('-') # make URL from author name
    return url

journalists['request_url'] = journalists.apply(lambda x: create_url(x['author_name_clean']), axis=1)
journalists['redirect_url'] = journalists.apply(lambda x: create_url(x['author_name_clean']), axis=1)

journalists

# add extra columns
for j in range(len(journalists)):
    journalists.loc[j, 'method'] = '-'
    journalists.loc[j, 'confidence'] = '-'

ua = UserAgent()

#### Experimenting with TF IDF Matching vs Levenshtein distance

lookup = list(muckrack_sitemap.url)
k_matches = 1
ngram_length = 3

def match_with_tfidf(wrong_url):
    wrong_url = [wrong_url]
    results = tm.matcher(wrong_url, lookup, k_matches, ngram_length)
    match_confidence = results['Match Confidence'].iloc[0]
    right_url = results['Lookup 1'].iloc[0]
    return right_url, match_confidence

def match_with_levenshtein(wrong_url):
    try:
        highest = process.extractOne(wrong_url, list(muckrack_sitemap.url))
        return highest
    except:
        return None

print("Computing 1 string match")

start = time.time()
for i in range(1):
    match_with_tfidf(journalists.iloc[i].request_url)
end = time.time()
delta = end - start
print("(TF-IDF) Seconds elapsed", delta)

start = time.time()
for i in range(1):
    match_with_levenshtein(journalists.iloc[i].request_url)
end = time.time()
delta = end - start
print("(Levenshtein) Seconds elapsed", delta)

print("Computing 10 string matches")

start = time.time()
for i in range(10):
    match_with_tfidf(journalists.iloc[i].request_url)
end = time.time()
delta = end - start
print("(TF-IDF) Seconds elapsed", delta)

start = time.time()
for i in range(10):
    match_with_levenshtein(journalists.iloc[i].request_url)
end = time.time()
delta = end - start
print("(Levenshtein) Seconds elapsed", delta)

print("Computing 100 string matches")

start = time.time()
for i in range(100):
    match_with_tfidf(journalists.iloc[i].request_url)
end = time.time()
delta = end - start
print("(TF-IDF) Seconds elapsed", delta)

start = time.time()
for i in range(100):
    match_with_levenshtein(journalists.iloc[i].request_url)
end = time.time()
delta = end - start
print("(Levenshtein) Seconds elapsed", delta)

#### Requesting URLs with Muckrack + TF IDF

journalists

for j in range(len(journalists)):
    url = journalists.iloc[j].request_url
    # print(j)
    try:
        req = Request(url, headers = {'User-Agent': ua.random})
        res = urlopen(req)
        sauce = res.read()
        journalists.loc[j, 'method'] = 'brute force'
        journalists.loc[j, 'redirect_url'] = res.url
        # print(j, "found brute force", url)
    except:
        url, confidence = match_with_tfidf(url)
        try:
            req = Request(url, headers = {'User-Agent': ua.random})
            res = urlopen(req)
            sauce = res.read()
            journalists.loc[j, 'method'] = 'tfidf match'
            journalists.loc[j, 'confidence'] = confidence
            journalists.loc[j, 'request_url'] = url
            journalists.loc[j, 'redirect_url'] = res.url
            # print(j, "found tfidf", url)
        except:
            pass
            # print(j, "found nothing", url)
            # journalists.loc[j, 'method'] = '-'
            # google scraper

#### Web search for remaining URLs

client = Algorithmia.client('simYm8DHcOm5sOaSIaAzH/Xzbin1')
algo = client.algo('specrom/Google_scraper/0.1.4')
algo.set_options(timeout=100) # optional

for i in range(len(journalists)):
    if (journalists.iloc[i].method == 'tfidf match' and float(journalists.iloc[i].confidence) < 0.85) or (journalists.iloc[i].method == '-') or (journalists.iloc[i].method == 'no match'):
        query = {"query": f"{journalists.iloc[i].author_name_clean} muckrack"}
        results = algo.pipe(query).result
        if results and 'muckrack' in results[0]['url']:
            journalists.at[i, 'request_url'] = results[0]['url']
            journalists.at[i, 'redirect_url'] = results[0]['url']
            journalists.at[i, 'method'] = 'google scrape'
            journalists.at[i, 'confidence'] = '-'    
        else:
            journalists.at[i, 'method'] = '-'

journalists.method.value_counts()

journalists = journalists[journalists.method != '-'].reset_index(drop = True)

### Extract additional information

def extract_data(soup):
    
    beats, title, location, links = '-', '-', '-', '-'
    try:
        if soup.find(class_ = 'person-details-item person-details-beats'):
            beats = ' '.join([line.strip() for line in soup.find(class_ = 'fa fa-fw fa-tasks icon-standard').findNext('div').text.split('\n') if line.strip() != ''])
    except:
        pass
    
    try:
        if soup.find(class_ = 'person-details-item person-details-title'):
            title = ' '.join([line.strip() for line in soup.find(class_ = 'fa fa-fw fa-building icon-standard').findNext('div').text.split('\n') if line.strip() != ''])
    except:
        pass
    
    try:
        if soup.find(class_ = 'person-details-item person-details-location'):
            location = ' '.join([line.strip() for line in soup.find(class_ = 'fa fa-fw fa-map-marker-alt icon-standard').findNext('div').text.split('\n') if line.strip() != ''])
    except:
        pass
    
    try: 
        if soup.find(class_ = 'profile-section-social'):
            urls = []
            for i in soup.find(class_ = 'profile-section-social'):
                temp = [j[5:] for j in str(i).split() if 'href' in j]
                if temp: 
                    temp = temp[0].strip('"')
                    urls.append(temp)
            links = ', '.join(urls)
    except:
        pass
    
    return beats, title, location, links

for i in range(len(journalists)):
    # print(i)
    muckrack = journalists.iloc[i].redirect_url
    req = Request(muckrack, headers = {'User-Agent': ua.random})
    try:
        res = urlopen(req)
        sauce = res.read()
        soup = bs.BeautifulSoup(sauce,'lxml')
        beats, title, location, links = extract_data(soup)
        journalists.loc[i, 'beats'], journalists.loc[i, 'title'], journalists.loc[i, 'location'], journalists.loc[i, 'links'] = beats, title, location, links
    except:
        journalists.loc[i, 'beats'], journalists.loc[i, 'title'], journalists.loc[i, 'location'], journalists.loc[i, 'links'] = '-', '-', '-', '-'
    # print(muckrack, extract_data(soup))

articles.dropna(subset = ['author_name_clean']).reset_index(drop = True).to_csv('../Data/articles_v1.csv', index = False)

journalists.dropna().reset_index(drop = True).to_csv('../Data/journalists_v1.csv', index = False)
