==> Notebooks/1.2 - Email matching.ipynb <==
### Loading libraries and files

import pickle
import pandas as pd
import json
import langdetect
import re
import Algorithmia
import bs4 as bs
import lxml
from urllib.request import Request, urlopen
import time
from fuzzywuzzy import process
import tfidf_matcher as tm
from fake_useragent import UserAgent
import numpy as np
import pandas as pd
import recordlinkage

# contains correct muckrack urls with either google scraping, tfidf matching or brute force approach.
author_df = pd.read_csv('../Data/journalists_v1.csv')

# original news dataset containing article info 
news_df = pd.read_csv('../Data/articles_v1.csv')

# emails dataset
emails_df = pd.read_csv("../Data/compression_test.txt.gz", compression='gzip',sep = '\t')
emails_df = emails_df.fillna("")
emails_df = emails_df[['email_address', 'first_name_df', 'second_name_df']]

author_df

news_df[['site_name']] = news_df[['site_name']].fillna(value = '-')
news_df

### Record deduplicating

# make a first name column for blocking using recordlinkage
def first_name(name):
    return name.split(' ')[0].lower()

# get the outlet the author most commonly works for and has the most articles for
def author_outlet(author):
    rows = news_df[news_df.author_name_clean == author]
    return rows.site_name.value_counts().index[0]

# concatenate above 2 fields to compare it with potential email matches
def concatenate_fields(author, outlet):
    return author + ' ' + outlet

def match_disambiguate_deduplicate(authors, emails, key):
    
    author_df['first_name_df'] = author_df.apply(lambda x: first_name(x['author_name_clean']), axis=1)
    author_df['common_outlet'] = author_df.apply(lambda x: author_outlet(x['author_name_clean']), axis=1)
    author_df['author + outlet'] = author_df.apply(lambda x: concatenate_fields(x['author_name_clean'], x['common_outlet']), axis=1)
       
    # create and indexer and block columns by the first name
    indexer = recordlinkage.Index()
    indexer.block('first_name_df')
    candidate_links = indexer.index(authors, emails)
    print("candidate_links:", len(candidate_links))
    
    # compare the author + outlet string from authors dataframe with the email string from the emails dataframe
    compare_cl = recordlinkage.Compare()
    compare_cl.string('author + outlet', 'email_address', label = 'match')
    features = compare_cl.compute(candidate_links, authors, emails)
    
    # find all unique author indexes for which a potential email match is found
    author_indexes = list(set([i[0] for i in features[features.match >= 0.5].index]))

    # to test, take only rows where the match confidence is >= 0.5
    features = features[features.match >= 0.5]
    
    # create an empty dataframe to store matches
    matches = pd.DataFrame (columns = ['author_index', 'email_index', 'match_threshold', 'scraper_response'])

    # for every author, having a variety of matching email indexes, take the one with highest match confidence
    for i in author_indexes:
        temp = features.loc[i].sort_values('match', ascending = False).head(1)
        row = pd.Series(list([i, temp.index[0], round(temp.match.values[0], 3), '-']), index = matches.columns)
        matches = matches.append(row, ignore_index = True) 
    
    # do a google search for every matching email IF they have a valid muckrack url found before, else we cannot confirm
    
    client = Algorithmia.client(key)
    algo = client.algo('specrom/Google_scraper/0.1.4')
    algo.set_options(timeout=300)
    
    for i in range(len(matches)):
        if i % 10 == 0: 
            print(i)
        if authors.iloc[int(matches.iloc[i].author_index)].method == 'no match':
            matches.loc[i, 'scraper_response'] = 'invalid'
            print("no valid muckrack url to compare with", matches.iloc[i].author_index)
        else:
            try:
                email_to_google = emails.iloc[int(matches.iloc[i].email_index)].email_address
                query = {"query": f"{email_to_google} muckrack"}
                results = algo.pipe(query).result
                matches.at[i, 'scraper_response'] = results
            except:
                pass
        
    return matches

email_matches = match_disambiguate_deduplicate(author_df, emails_df, 'simwGAeoL9oH26JJvCS0pBAGzcD1')

email_matches

email_matches

### Validating match quality

def found_match(author_index, email_index, scraper_response):
    if not scraper_response or scraper_response == 'invalid' or scraper_response == '-':
        return 'Invalid'
    else:
        # print(author_df.iloc[author_index].redirect_url, scraper_response[0]['url'])
        # print()
        return author_df.iloc[author_index].redirect_url == scraper_response[0]['url']

email_matches['email_checks_out'] = email_matches.apply(lambda x: found_match(x['author_index'], x['email_index'], x['scraper_response']), axis=1)

email_matches[email_matches.email_checks_out == True].match_threshold.mean()

email_matches[email_matches.email_checks_out == False].match_threshold.mean()

email_matches.email_checks_out.value_counts()

email_matches[email_matches.email_checks_out == True]

# example
author_df.iloc[21].author_name_clean, author_df.iloc[21].common_outlet, emails_df.iloc[617881]

email_matches[email_matches.email_checks_out == 'Invalid'].match_threshold.mean()

email_matches.reset_index(drop = True).to_csv('../Data/email_matches.csv', index = False)

author_df['email'] = np.nan
author_df

for i in range(len(email_matches)):
    cell = email_matches.iloc[i]
#     print(cell.email_checks_out)
    if cell.email_checks_out == 'True':
#         print(emails_df.loc[cell.email_index, 'email_address'])
        author_df.loc[cell.author_index, 'email'] = emails_df.loc[cell.email_index, 'email_address']

author_df.email.value_counts()

author_df

author_df.to_csv('../Data/journalists_v2.csv', index = False)
