==> Notebooks/1.3 - Collecting outlet data.ipynb <==
### Load libraries and files

import pickle
import pandas as pd
import json
import langdetect
import re
import Algorithmia
import bs4 as bs
import lxml
from urllib.request import Request, urlopen
import time
from fuzzywuzzy import process
import tfidf_matcher as tm
from fake_useragent import UserAgent
import numpy as np
import pandas as pd
import recordlinkage

articles = pd.read_csv('../Data/articles_v1.csv')

articles

### Collecting outlet Muckrack URLs

outlets = pd.DataFrame(articles.groupby('site_name')['author_name_clean'].nunique()).reset_index().rename(columns = {'author_name_clean': 'unique_journalist_count'})

outlets

len(articles[articles.site_name == 'seattlepi.com'].author_name_clean.unique())

len(articles[articles.site_name == 'AZFamily'].author_name_clean.unique())

outlets['muckrack_url'] = np.nan

outlets

client = Algorithmia.client('simKUmpUiy7PIeS1Sh7EV/cBoMJ1')
algo = client.algo('specrom/Google_scraper/0.1.4')
algo.set_options(timeout=300) # optional

# for i in range(len(outlets)):
#     query = {"query": f'"media-outlet" site: muckrack {outlets.iloc[i].site_name}'}
#     # print(query)
#     try:
#         res = algo.pipe(query).result[0]['url']
#         outlets.loc[i, 'muckrack_url'] = res
#     except:
#         pass
# #     else:
# #         print('none')

### Collecting outlet website URLs

outlets['website_url'] = np.nan

outlets

ua = UserAgent()
for i in range(len(outlets)):
    print(i)
    url = outlets.iloc[i].muckrack_url
    try:
        req = Request(url, headers = {'User-Agent': ua.random})
        sauce = urlopen(req).read()
        soup = bs.BeautifulSoup(sauce,'lxml')
        link = soup.find_all(class_ = 'mr-contact-item-inner')[0].text
        outlets.loc[i, 'website_url'] = link
    except:
        pass

outlets.website_url.value_counts()

### Collecting outlet Alexa rankings and popularity

outlets['popularity'] = 0
outlets['reach_rank'] = 0
outlets['rank_delta'] = 0
outlets['country_rank'] = 0

outlets

for i in range(len(outlets)):
    
    url = outlets.iloc[i].website_url
    
    try:
        req = Request('http://data.alexa.com/data?cli=10&dat=s&url={}'.format(url), headers = {'User-Agent': ua.random})
        sauce = urlopen(req).read()
        soup = bs.BeautifulSoup(sauce,'xml')

        try:
            popularity = int(soup.find('POPULARITY')['TEXT'])
            outlets.loc[i, 'popularity'] = popularity
        except:
            pass

        try: 
            reach_rank = int(soup.find('REACH')['RANK'])
            outlets.loc[i, 'reach_rank'] = reach_rank
        except:
            pass

        try: 
            rank_delta = int(soup.find('RANK')['DELTA'])
            outlets.loc[i, 'rank_delta'] = rank_delta
        except:
            pass

        try:
            country_rank = int(soup.find('COUNTRY')['RANK'])
            outlets.loc[i, 'country_rank'] = country_rank
        except:
            pass

    except:
        pass

# print(popularity, reach_rank, rank_delta, country_rank)

outlets.describe()

outlets.to_csv('../Data/outlets_v1.csv', index = False)
