==> Notebooks/2 - Preprocessing data.ipynb <==
### Load files and libraries

import pandas as pd
import ftfy
import random
from fake_useragent import UserAgent
import nltk
import pickle
import pandas as pd
from nltk.corpus import wordnet
import json
from afinn import Afinn
import plotly.express as px
from nltk.corpus import stopwords
from nltk.stem.wordnet import WordNetLemmatizer
import string
import re
import ftfy
import humanize
import datetime
import langdetect

articles = pd.read_csv('../Data/articles_v1.csv')

articles

### Cleaning and preprocessing text columns

stop = set(stopwords.words('english'))
exclude = set(string.punctuation)
lemma = WordNetLemmatizer()
tag_dict = {"J": wordnet.ADJ, "N": wordnet.NOUN, "V": wordnet.VERB, "R": wordnet.ADV}

# Cleaning the text sentences so that punctuation marks, stop words &amp; digits are removed
# Words are lemmatized according to their POS tags
def clean(doc):
    doc = ftfy.fix_text(doc)
    stop_free = " ".join([i for i in doc.lower().split() if i not in stop])
    punc_free = ''.join(ch for ch in stop_free if ch not in exclude)
    normalized = ""
    for word in punc_free.split():
        tag = nltk.pos_tag([word])[0][1][0].upper()
        normalized = normalized + " " + str(lemma.lemmatize(word, tag_dict.get(tag, wordnet.NOUN)))
    processed = re.sub(r"\d+","", normalized)
    y = processed.split()
    return ' '.join(y)

### Cleaning article text (title x4, description x2, full_text x1)

articles = articles.fillna('-')

articles['content'] = ((articles['title'] + ' ') * 4) + ((articles['description'] + ' ') * 2) + (articles['full_text'])
articles['content'] = articles.apply(lambda x: clean(x['content']), axis = 1)

articles

### Calculating Afinn sentiment for all articles

afinn = Afinn(language='en')
def get_article_sentiment(text):
    return round(afinn.score(text) / len(text.split()) * 100, 2)

articles['article_sentiment'] = articles.apply(lambda x: get_article_sentiment(x['full_text']), axis=1)

articles

articles.describe()

articles.to_csv('../Data/articles_v2.csv')

### Cleaing IAB classifier text

iab_text = pd.read_csv('../Data/IAB_text_v1.csv')

iab_text['with_bs4_clean'] = iab_text.apply(lambda x: clean(x['with_bs4']), axis=1)

iab_text['with_justext_clean'] = iab_text.apply(lambda x: clean(x['with_justext']), axis=1)

iab_text

# iab_text.to_csv('../Data/IAB_text_v2.csv', index = False)

iab_text = pd.read_csv('../Data/IAB_text_v2.csv')

regex_pattern = re.compile(pattern = "["
        u"\U0001F600-\U0001F64F"  # emoticons
        u"\U0001F300-\U0001F5FF"  # symbols & pictographs
        u"\U0001F680-\U0001F6FF"  # transport & map symbols
        u"\U0001F1E0-\U0001F1FF"  # flags (iOS)
                           "]+", flags = re.UNICODE)

def remove_emojis(text):
    try:
        return regex_pattern.sub(r'',text)
    except:
        return np.nan

iab_text['with_bs4_clean'] = iab_text.apply(lambda x: remove_emojis(x['with_bs4_clean']), axis = 1)

iab_text['with_justext_clean'] = iab_text.apply(lambda x: remove_emojis(x['with_justext_clean']), axis = 1)

for i in range(len(iab_text)):
    try:
        text = iab_text.iloc[i].with_bs4
        lang = langdetect.detect(text)
        if lang != 'en':
            iab_text.loc[i, 'with_bs4_clean'] = np.nan
    except Exception as e:
        iab_text.loc[i, 'with_bs4_clean'] = np.nan

iab_text.dropna(subset = ['with_bs4_clean'])

for i in range(len(iab_text)):
    try:
        text = iab_text.iloc[i].with_justext
        lang = langdetect.detect(text)
        if lang != 'en':
            iab_text.loc[i, 'with_justext_clean'] = np.nan
    except Exception as e:
        iab_text.loc[i, 'with_justext_clean'] = np.nan

iab_text.dropna(subset = ['with_justext_clean'])

iab_text.to_csv("../Data/IAB_text_v3.csv", index = False)
